A long-wanted cleanup of message classification.
Pulled all message classification out into a testable struct with
a couple of functions on it. Retorts fall into a couple of patterns
that are now expressed via an enum.  This reduces the Loudbot struct
to the glue between Slack inputs and redis-stored outputs.

Wrote some more tests.

Because I can't build x86 targets on my M1 laptop, it's time to
switch to using a Github action to build releases. Thanks, Uncle Satya,
for the free CPU. Goodbye Makefile; greetings justfile.

Bumped the Rust edition to 2021.