Refactoring in prep for generalization (#5)
I want to be able to add both a Discord integration and a Slack RTM
integration, so I refactored to pull all the bits that know about
Slack into a single struct. This struct now has a loudbot brain, which
in turn manages the redis connection and the input classifier. The
LOUDBOT executable is an axum app that constructs both the brain and
the slack integration from env var config, and then wires up
endpoints. I haven't yet found an abstraction for the integrations
that is worth the effort, and I do not expect to find it until I have
all three of the planned clients working.

The next step here is to resurrect the RTM version of the integration
from the branch where it lives right now. I might have to do something
to break up the project into multiple crates at that point.